refactor: shift stack & queue
